+++ fast_ui/example/lib/examples/fast_rx_example.dart
@@ -1,4 +1,4 @@
-import 'package:fast_rx/fast_rx.dart';
+import 'package:fast_ui/fast_ui.dart';
 import 'package:flutter/material.dart';
 
 class FastRxExample extends StatelessWidget {
@@ -58,8 +58,7 @@ class FastRxExample extends StatelessWidget {
             const SizedBox(height: 32),
             FastBuilder(
               () => ElevatedButton(
-                child:
-                    Text((_rebuild.value ? 'Disable' : 'Enable') + ' rebuild'),
+                child: Text('${_rebuild.value ? 'Disable' : 'Enable'} rebuild'),
                 onPressed: () => _rebuild.value = !_rebuild.value,
               ),
             ),