+++ fast_ui/lib/fast_ui.dart
@@ -5,3 +5,4 @@ export 'package:fast_extensions/fast_extensions.dart';
 export 'package:fast_nav/fast_nav.dart';
 export 'package:fast_overlays/fast_overlays.dart';
 export 'package:fast_rx/fast_rx.dart';
+export 'package:fast_rx_persistence/fast_rx_persistence.dart';